refactor: folder name utl to utils
