+++ src/modules/auth/ForgotPassword/ForgotPasswordView.tsx
@@ -6,7 +6,7 @@ import { Footer, Link } from "modules/auth/ForgotPassword/styles";
 import { IForgotPasswordView } from "modules/auth/ForgotPassword/types";
 import { defineTitleMessage } from "modules/auth/ForgotPassword/utils/define-title-message";
 import { defineSubtitleMessage } from "modules/auth/ForgotPassword/utils/define-subtitle-message";
-import { fieldErrors } from "shared/util/fieldErrors";
+import { fieldErrors } from "shared/utils/fieldErrors";
 
 const ForgotPasswordView: React.FC<IForgotPasswordView> = props => {
 	const {
