+++ src/modules/auth/SignUp/SignUpView.tsx
@@ -6,7 +6,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import Greeting from "shared/components/Greeting";
 import TextFieldPassword from "shared/components/TextFieldPassword";
-import { fieldErrors } from "shared/util/fieldErrors";
+import { fieldErrors } from "shared/utils/fieldErrors";
 
 const SignUpView: React.FC<ISignUpView> = props => {
 	const { register, errors, handleSubmit, signUpSubmit, isLoading } = props;
