+++ src/modules/texts/Text/components/CardText/components/CardHeader/CardHeader.tsx
@@ -5,7 +5,7 @@ import { ICON_SIZE } from "shared/constants";
 import { Link } from "react-router-dom";
 import { useDialogText } from "modules/texts/Text/shared/hooks/useDialogText";
 import React from "react";
-import { createURLQueryParams } from "shared/util/createURLQueryParams";
+import { createURLQueryParams } from "shared/utils/createURLQueryParams";
 import { ICardHeader } from "modules/texts/Text/components/CardText/components/CardHeader/types";
 
 const CardHeader: React.FC<ICardHeader> = ({ text_id, board_id }) => {
