+++ src/shared/utils/test/renderRequiredAuth.tsx
